pyproject.toml
[project]
name = "openavc-simulator"
version = "0.1.0"
description = "AV device simulator for OpenAVC — simulates projectors, displays, switchers, DSPs, cameras, and more"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pyyaml>=6.0",
    "websockets>=12.0",
    "aiohttp>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["simulator*"]

[project.scripts]
openavc-simulator = "simulator.__main__:main"
